+++ docs/changelog.rst
@@ -5,6 +5,10 @@ CHANGELOG
 2.126.0+dev     (XXXX-XX-XX)
 ----------------------------
 
+**Improvements**
+
+* Remove deleted attribut of blade model
+
 
 2.126.0         (2026-08-06)
 ----------------------------